Fixed #24: Purging a specific empty folder does nothing.
Turns out none of the connecting glue between the event required to purge a folder and the code that does the heavy lifting was even in existence.

I carefully made sure the code flow, method locations and names paralleled the code for emptying the trash. Smoother debugging that way.

I also found that CachedUserInfo::PurgeFolder was misnamed as it only purged the contents of the folder. I corrected that and then added a new method that re-used the old name that actually does delete the folder and its contents.

It seems the client has some cruft as it makes some interesting calls:
* When Emptying the trash, HandlePurgeInventoryDescendents() is called.  This makes sense as the trash folder should be left unmolested.
* When Purging an empty folder located in the Trash, HandlePurgeInventoryFolder() is called.  This is what used to be RemoveInventoryFolder().  This also makes sense as the folder and all of its contents, of which there are none in this case, needs to be emptied.
* When Purging a folder that has children, both HandlePurgeInventoryFolder() and HandlePurgeInventoryDescendents() are called - in that order.  This make little sense as it should have been the same case as the above: only HandlePurgeInventoryFolder(). I suspect this is cruft from back when LL may have used the viewer to keep the DB consistent...

That last shouldn't be a big issue: HandlePurgeInventoryFolder() will remove the whole shooting match and the subsequent HandlePurgeInventoryDescendents() call will find that there's nothign to purge and simply return. Though it might have a null pointer problem when it tries to use the null result of GetFolderAttributes()...  Need to fix the Empty Trash bug in order to test.

+++ OpenSim/Region/Framework/Scenes/Scene.PacketHandlers.cs
@@ -671,11 +671,61 @@ namespace OpenSim.Region.Framework.Scenes
 
                 m_log.InfoFormat("[AGENT INVENTORY]: Purging descendents of {0} {1} for user {2}", folderID, folder.Name, remoteClient.Name);
 
-                if (!userProfile.PurgeFolder(folder))
+                if (!userProfile.PurgeFolderContents(folder))
                 {
                     m_log.ErrorFormat(
                             "[AGENT INVENTORY]: Failed to purge folder for user {0} {1}",
                             remoteClient.Name, remoteClient.AgentId);
+                }
+            });
+        }
+
+        /// <summary>
+        /// This should delete the given directory and all its descendents recursively.
+        /// </summary>
+        /// <param name="remoteClient"></param>
+        /// <param name="folderID"></param>
+        public void HandlePurgeInventoryFolder(IClientAPI remoteClient, UUID folderID)
+        {
+            Util.FireAndForget(Util.PoolSelection.LongIO, delegate(object obj)
+            {
+                CachedUserInfo userProfile = CommsManager.UserProfileCacheService.GetUserDetails(remoteClient.AgentId);
+
+                if (null == userProfile)
+                {
+                    m_log.ErrorFormat(
+                        "[AGENT INVENTORY]: Could not find user profile for {0} {1}",
+                        remoteClient.Name, remoteClient.AgentId);
+                    return;
+                }
+
+                //make sure the user owns the source folder
+                InventoryFolderBase folder = userProfile.GetFolderAttributes(folderID);
+
+                if (folder.Owner != remoteClient.AgentId)
+                {
+                    m_log.ErrorFormat(
+                        "[AGENT INVENTORY]: Not purging descendents of {0} for user {1}, user does not own the folder",
+                        folderID, remoteClient.Name);
+
+                    return;
+                }
+
+                //make sure the folder is either the lost and found, the trash, or a descendant of it
+                if (!CheckFolderHeirarchyIsAppropriateForPurge(folder, userProfile))
+                {
+                    m_log.ErrorFormat(
+                        "[AGENT INVENTORY]: Not purging descendents of {0} for user {1}, folder is not part of a purgeable heirarchy",
+                        folderID, remoteClient.Name);
+                }
+
+                m_log.InfoFormat("[AGENT INVENTORY]: Purging {0} {1} for user {2}", folderID, folder.Name, remoteClient.Name);
+
+                if (!userProfile.PurgeFolder(folder))
+                {
+                    m_log.ErrorFormat(
+                        "[AGENT INVENTORY]: Failed to purge folder for user {0} {1}",
+                        remoteClient.Name, remoteClient.AgentId);
                 }
             });
         }
